+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make `includes` use glob patterns by @daimond113
 - Use symlinks for workspace dependencies to not require reinstalling by @daimond113
 
+### Fixed
+- Install dependencies of packages in `x` command
+
 ## [0.5.0-rc.12] - 2024-11-22
 ### Added
 - Support fallback Wally registries by @daimond113